Move Power Endpoints to Router
This resolves #12

This change moves the power data endpoints to a
router to match the pattern established for user
endpoints.
